migrated layer building scripts, updated docs, and made VM builder code changes
