+++ README.md
@@ -24,48 +24,49 @@ srm version
 srm status
 ```
 
-2. Update your config file in your home directory.  At minimum, you'll need:
-  * Your own Lambda Execution Role 'arn' value.
-  * An S3 bucket and an S3 bucket key (folder).
-  * Lambda function name
+2. At minimum for publishing function code, you'll need the following filled out in your ~/serveRmore.yaml file:
+```
+aws:
+  s3_bucket:
+  s3_key:
+lambda:
+  arn_role: arn:aws:iam::<AWS_ID>:role/lambda_basic_execution
+  arn_runtime_layer: arn:aws:lambda:us-east-1:<AWS_ID>:layer:<name>:<version>
+  arn_custom_layer: arn:aws:lambda:us-east-1:<AWS_ID:layer:<name>:<version>
+  name:
+  r_version:
+  zip_file_name:
+```
 
 3. Create a new lambda.R script and create a "handler" method in R.  Insert "hello world" or custom code inside your handler method.
 
-4. Create a new package.R script that will zip up your lambda.R and place on your AWS S3 bucket and key.
+4. Create a new deploy.R script that will zip up your lambda.R and place on your AWS S3 bucket and key.
 
 5. To deploy your zip file directly to Lambda, try out our new workflow here.
 ```
-srm lambda init
 srm lambda create
 srm lambda update
 srm lambda invoke
 srm lambda destroy
 ```
 
-NOTE: 'init' will establish your R Runtime ARN value in your config.  'create' will establish a brand new Lambda function if it does not exist, and publish your zip file.  'update' will republish your zip file, if your lambda function already exists.  
+NOTE: 'create' will establish a brand new Lambda function if it does not exist, and publish your zip file.  'update' will republish your zip file, if your lambda function already exists.  
 
 ## Custom R Runtime Layers
 
 To use additional R custom runtime layers that are prebuilt, or to help you build your own, refer to this guide: [R Runtime Custom Layers](custom_layer/README.md)
 
-We have an older way of deploying on Lambda that we still support for the time being.  You can learn more here: [R Environment in Python Runtime](deprecated.md)
-
 ## Roadmap
 
-v0.1.0
+v0.1.1
 * Create more automation on custom layer building
-    - Revise ServeRmore.yaml config file for new configurations
-    - Launch a temporary EC2 VM
-    - Pull down the github repo from bakdata
     - Introduce custom layer settings from config file
     - Build & publish new layer
     - Save the ARN of the new layer to config
-    - Terminate the VM
-* Remove old EC2 Virtual Machine on Python runtime workflow
 
 ## Changelog
 
-v0.0.3 - Removed original and deprecated deployment steps for custom Lambda packages that used r2py, Python 2, and R 3.4.2.  
+v0.1.0 - Removed original and deprecated deployment steps for custom Lambda packages that used r2py, Python 2, and R 3.4.2. Added new lambda layer builder VM commands.  Also migrated scripts for R, R-runtime, and R-gbm layers into subfolder.
 
 v0.0.2 - Introduces running R directly on Lambda via AWS' new custom layers feature, big thanks to [@bakdata](https://github.com/bakdata).
 * Several new commands for initiating the R runtime version, creating new functions, updating them, and destroying them.
