+++ VERSION.py
@@ -1,0 +1,3 @@
+#!python3
+
+VERSION = "0.1.0"
