+++ layers/r-gbm/compile.sh
@@ -1,0 +1,26 @@
+#!/bin/bash
+
+set -euo pipefail
+
+if [[ -z ${1+x} ]];
+then
+    echo 'version number required'
+    exit 1
+else
+    VERSION=$1
+fi
+
+BASE_DIR=$(pwd)
+BUILD_DIR=${BASE_DIR}/build/
+R_DIR=/opt/R/
+
+rm -rf ${BUILD_DIR}
+
+export R_LIBS=${BUILD_DIR}/layer/R/library
+mkdir -p ${R_LIBS}
+${R_DIR}/bin/Rscript -e 'install.packages("Matrix", repos="http://cran.r-project.org")'
+${R_DIR}/bin/Rscript -e 'install.packages("survival", repos="http://cran.r-project.org")'
+${R_DIR}/bin/Rscript -e 'install.packages("gridExtra", repos="http://cran.r-project.org")'
+${R_DIR}/bin/Rscript -e 'install.packages("lattice", repos="http://cran.r-project.org")'
+${R_DIR}/bin/Rscript -e 'install.packages("gbm", repos="http://cran.r-project.org")'
+chmod -R 755 ${BUILD_DIR}/layer/
