+++ layers/r-runtime/src/bootstrap
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+set -euo pipefail
+
+/opt/R/bin/Rscript /opt/bootstrap.R
