+++ srvl_config.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
 
-import ctypes, os, io, yaml
+import ctypes, os, io, yaml, time, paramiko, VERSION
 from pathlib import Path
+from setuptools import setup
 
 class srvlConfig:
 
@@ -9,7 +10,6 @@ class srvlConfig:
         self.file = str(Path.home())+"/serveRmore.yaml"
         self.settings = {}
         self.pwd = os.getcwd()
-        self.print_msg = 0
         os.chdir(str(Path.home())+"/")
         self.load_all()
 
@@ -20,7 +20,7 @@ class srvlConfig:
             print(yaml.dump(self.load(), default_flow_style=False))
 
     def version(self):
-        print("ServeRmore 0.0.3")
+        print("ServeRmore "+VERSION.VERSION)
 
     def load_all(self):
         self.settings = self.load()
@@ -55,6 +55,13 @@ class srvlConfig:
         with open(self.file, 'r') as stream:
             return yaml.load(stream, Loader=yaml.FullLoader)
 
+
+    def load(self):
+        if not self.exists():
+            raise ValueError("serveRmore.yaml configuration can not be found in your home directory.  Please refer to ServeRmore README instructions.")
+        with open(self.file, 'r') as stream:
+            return yaml.load(stream, Loader=yaml.FullLoader)
+
     def replace(self):
         if self.exists():
             os.replace(self.file, self.file + ".old")
@@ -67,7 +74,6 @@ class srvlConfig:
 
     def help(self):
         s = """\nPlease use the following commands for AWS:
-
     bash$: srm help | settings | version
         - utility helpers
     bash$: srm lambda init | list
@@ -75,35 +81,78 @@ class srvlConfig:
     bash$: srm lambda create | update | destroy
         - Create your function, update it, or destroy it
     bash$: srm lambda invoke
-        - Run your function from the command line\n\n"""
+        - Run your function from the command line
+    bash$: srm create | ssh | cpu | status | destroy
+        - Create and interact with your layer builder VM\n\n"""
         print(s)
 
-    def reset(self):
-        return yaml.load("""
-        aws:
-            s3_bucket:
-            s3_key:
-            default_security_group: 
-            mount_efs: 
-            private_key: 
-            ssh_security_group: 
-            subnet: 
-        lambda:
-            name:
-            r_version: 3.5.3
-            arn_role:
-            arn_runtime_layer:
-            arn_custom_layer:
-            zip_file_name: lambda.zip
-        rs: 
-            ami: 
-            cran_packages: 
-            domain_name:
-            ec2_mesgs: 
-            ec2_types:
-            instance_id: 
-            instance_type: 
-            version: 3.5.3
-        git: 
-            private_key: github.pem
-        """)
+class srvlConnect:
+
+    def evoke_ssh(self, user_name, domain_name, remote_port=22):
+        import subprocess
+        import sys
+        COMMAND="ssh "+ user_name + "@" + domain_name + " -o TCPKeepAlive=yes -o ServerAliveInterval=50 -p " + str(remote_port)
+        result = subprocess.call(COMMAND,shell=True)
+
+    def evoke_multitunnel(self, tunnel_str, connect_str, private_key):
+        import subprocess
+        import sys
+        COMMAND="ssh -i ~/.ssh/"+private_key+" -F ~/.ssh/config -N " + tunnel_str + " " + connect_str + " -o TCPKeepAlive=yes -o ServerAliveInterval=50 -o ExitOnForwardFailure=yes"
+        print("SSH Tunnel Connected - please type Ctrl-C to terminate.")
+        result = subprocess.call(COMMAND,shell=True)
+
+    def initiate_ssh(self,user,key,domain):
+        self.user_name = user
+        self.private_key = key
+        self.domain_name = domain
+        import paramiko
+        from pathlib import Path
+        self.client = paramiko.SSHClient()
+        self.client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+        self.client.load_system_host_keys()
+        key_object = paramiko.RSAKey.from_private_key_file(str(Path.home())+"/.ssh/"+self.private_key)
+        proceed = False
+        while proceed == False:
+            print("Attempting SSH connection...")
+            try:
+                self.client.connect(hostname=self.domain_name, port=22, username=self.user_name, pkey=key_object)
+                proceed = True
+                print("SSH successful!")
+            except Exception as e:
+                proceed = False
+                print("SSH failed... temporary delay enabled.")
+                time.sleep(5)
+        self.session = self.client.get_transport().open_session()
+
+    def run_ssh(self,cmd):
+        stdin, stdout, stderr = self.client.exec_command(cmd)
+        self.wait_for_command(stdout)
+
+    def upload_file_ssh(self, local_path, remote_path, value):
+        sftp = self.client.open_sftp()
+        print("copying from: "+local_path+value)
+        print("copying to: "+remote_path+value)
+        sftp.put(local_path+value, remote_path+value)
+
+    def terminate_ssh(self):
+        self.client.close()
+
+    def wait_for_command(self,stdout):
+        import select
+        while not stdout.channel.exit_status_ready():
+            if stdout.channel.recv_ready():
+                rl, wl, xl = select.select([stdout.channel], [], [], 0.0)
+                if len(rl) > 0:
+                    print(str(stdout.channel.recv(1024), 'utf-8').replace('\\n', '\n'))
+
+    def get_aws_access_key(self):
+        import botocore.session
+        return botocore.session.get_session().get_credentials().access_key
+
+    def get_aws_secret_key(self):
+        import botocore.session
+        return botocore.session.get_session().get_credentials().secret_key
+
+    def get_aws_region(self):
+        import botocore.session
+        return botocore.session.get_session().get_config_variable('region')
